feat(configprovider): migrate locale type to useLocale
