+++ src/components/locale/useLocale.tsx
@@ -1,8 +1,69 @@
-import { useContext, useMemo } from 'react';
+import { createContext, useContext, useMemo } from 'react';
 
-import type { Locale, LocaleContextProps } from '../configProvider';
-import { LocaleContext } from '../configProvider';
 import defaultLocaleData from './zh-CN';
+
+export interface Locale {
+    locale: string;
+    BlockHeader?: { expand: string; collapse: string };
+    ChromeDownload?: { description: string; download: string };
+    CopyIcon?: { copied: string; notSupport: string; copy: string };
+    EditCell?: { complete: string; cancel: string; modify: string };
+    EditInput?: { description: string };
+    Fullscreen?: { exitFull: string; full: string };
+    GlobalLoading?: {
+        loading: string;
+    };
+    GoBack?: {
+        back: string;
+    };
+    LoadError?: {
+        please: string;
+        get: string;
+        refresh: string;
+        title: string;
+    };
+    ModalWithForm?: {
+        okText: string;
+        cancelText: string;
+    };
+    MulSelectDropdown?: {
+        open: string;
+        cancelText: string;
+        okText: string;
+        selectAll: string;
+    };
+    MultiSearchInput?: {
+        case: string;
+        extract: string;
+        head: string;
+        tail: string;
+    };
+    MxGraph?: { newNode: string };
+    NotFound?: {
+        description: string;
+    };
+    ProgressLine?: {
+        description: string;
+    };
+    RenderFormItem?: {
+        description: string;
+    };
+    SearchModal?: {
+        title: string;
+        placeholder: string;
+    };
+    SpreadSheet?: {
+        description: string;
+        copy: string;
+        copyAll: string;
+    };
+}
+
+export interface LocaleContextProps {
+    locale: Locale;
+}
+
+export const LocaleContext = createContext<LocaleContextProps>(undefined);
 
 export type LocaleComponentName = keyof Locale;
